WIP: refactored themes to support live-updating
Key changes:
  - app/coder/style.py is gone, moved into app/themes/_themes.py
  - Setting the theme (and loading the theme specs) is done only at
    the app level. So theme-setting wx events should bind to there
  - after loadng the theme in the app we then call _applAppTheme()
    for each object
  - _applAppTheme is a method of ThemeMixin. By default, it recursively
    sets background/foreground to frame_bg and txt_default
  - objects with more complex needs (multiple backgrounds for different
    children) need to provide a `_applAppTheme()` method
