+++ psychopy/app/_psychopyApp.py
@@ -15,12 +15,15 @@ profiling = False  # turning on will save profile files in currDir
 import sys
 import platform
 import psychopy
+from psychopy import prefs
 from pkg_resources import parse_version
 from psychopy.constants import PY3
-import io
 from . import urls
 from . import frametracker
-from .themes import ThemeMixin
+from . import themes
+
+import io
+import json
 
 if not hasattr(sys, 'frozen'):
     try:
@@ -43,7 +46,7 @@ from psychopy.localization import _translate
 
 # needed by splash screen for the path to resources/psychopySplash.png
 import ctypes
-from psychopy import preferences, logging, __version__
+from psychopy import logging, __version__
 from psychopy import projects
 from . import connections
 from .utils import FileDropTarget
@@ -64,8 +67,8 @@ travisCI = bool(str(os.environ.get('TRAVIS')).lower() == 'true')
 # Enable high-dpi support if on Windows. This fixes blurry text rendering.
 if sys.platform == 'win32':
     # get the preference for high DPI
-    if 'highDPI' in preferences.prefs.app.keys():  # check if we have the option
-        enableHighDPI = preferences.prefs.app['highDPI']
+    if 'highDPI' in psychopy.prefs.app.keys():  # check if we have the option
+        enableHighDPI = psychopy.prefs.app['highDPI']
 
         # check if we have OS support for it
         if enableHighDPI:
@@ -76,11 +79,11 @@ if sys.platform == 'win32':
                     "High DPI support is not appear to be supported by this version"
                     " of Windows. Disabling in preferences.")
 
-                preferences.prefs.app['highDPI'] = False
-                preferences.prefs.saveUserPrefs()
+                psychopy.prefs.app['highDPI'] = False
+                psychopy.prefs.saveUserPrefs()
 
 
-class MenuFrame(wx.Frame, ThemeMixin):
+class MenuFrame(wx.Frame, themes.ThemeMixin):
     """A simple empty frame with a menubar, should be last frame closed on mac
     """
 
@@ -155,7 +158,7 @@ class _Showgui_Hack(object):
         core.shellCall([sys.executable, noopPath])
 
 
-class PsychoPyApp(wx.App, ThemeMixin):
+class PsychoPyApp(wx.App, themes.ThemeMixin):
 
     def __init__(self, arg=0, testMode=False, **kwargs):
         """With a wx.App some things get done here, before App.__init__
@@ -173,6 +176,7 @@ class PsychoPyApp(wx.App, ThemeMixin):
         self.version = psychopy.__version__
         # set default paths and prefs
         self.prefs = psychopy.prefs
+        self._currentThemeSpec = None
 
         self.keys = self.prefs.keys
         self.prefs.pageCurrent = 0  # track last-viewed page, can return there
@@ -320,7 +324,7 @@ class PsychoPyApp(wx.App, ThemeMixin):
                        float(wx.GetDisplaySizeMM()[0]) * 25.4)
         if not (50 < self.dpi < 120):
             self.dpi = 80  # dpi was unreasonable, make one up
-        self.theme = prefs.app['theme']
+        self.theme = self.prefs.app['theme']
 
         if sys.platform == 'win32':
             # wx.SYS_DEFAULT_GUI_FONT is default GUI font in Win32
@@ -608,11 +612,11 @@ class PsychoPyApp(wx.App, ThemeMixin):
         # have to reimport because it is only local to __init__ so far
         from .runner.runner import RunnerFrame
         title = "PsychoPy3 Experiment Runner (v{})".format(self.version)
-        runner = RunnerFrame(parent=None,
+        self.runner = RunnerFrame(parent=None,
                              id=-1,
                              title=title,
                              app=self)
-        return runner
+        return self.runner
 
     def OnDrop(self, x, y, files):
         """Not clear this method ever gets called!"""
@@ -877,6 +881,59 @@ class PsychoPyApp(wx.App, ThemeMixin):
         idle.doIdleTasks(app=self)
         evt.Skip()
 
+    def onThemeChange(self, event):
+        """Handles a theme change event (from a window with a themesMenu)"""
+        win = event.EventObject.Window
+        newTheme = win.themesMenu.FindItemById(event.GetId()).Label
+        self.theme = newTheme
+
+    def _applyAppTheme(self):
+        """Overrides ThemeMixin for this class"""
+        for frameRef in self._allFrames:
+            frame = frameRef()
+            if hasattr(frame, '_applyAppTheme'):
+                frame._applyAppTheme()
+
+    @property
+    def theme(self):
+        """The theme to be used through the application"""
+        return prefs.app['theme']
+
+    @theme.setter
+    def theme(self, value):
+        """The theme to be used through the application"""
+        try:
+            with open("{}//{}.json".format(prefs.paths['themes'], value), "rb") as fp:
+                spec = json.load(fp)
+        except:
+            with open("{}//{}.json".format(prefs.paths['themes'], "PsychopyLight"), "rb") as fp:
+                spec = json.load(fp)
+
+        # Check that minimum spec is defined
+        if 'base' in spec:
+            base = spec['base']
+            if not (
+                all(key in base for key in ['bg', 'fg', 'font'])
+            ):
+                return
+        else:
+            return
+
+        # we have a valid theme so continue
+        self._currentThemeSpec = spec
+        themes.ThemeMixin.codeColors = spec  # class attribute for all mixin subclasses
+        themes.ThemeMixin.mode = spec['app'] if 'app' in spec else 'light'
+        themes.ThemeMixin.icons = spec['icons'] if 'icons' in spec else 'modern'
+        if spec['app'] == 'dark':
+            themes.ThemeMixin.appColors = themes.cs_dark
+        elif spec['app'] == 'light':
+            themes.ThemeMixin.appColors = themes.cs_light
+
+        # Override base font with user spec if present
+        key = 'outputFont' if isinstance(self, wx.py.shell.Shell) else 'codeFont'
+        if prefs.coder[key] != "From theme...":
+            base['font'] = prefs.coder[key]
+        self._applyAppTheme()
 
 if __name__ == '__main__':
     # never run; stopped earlier at cannot do relative import in a non-package
