+++ psychopy/app/coder/sourceTree.py
@@ -9,7 +9,7 @@
 
 from __future__ import absolute_import, print_function
 from collections import deque
-from ..style import cs, cLib
+from ..themes import ThemeMixin
 from psychopy.app.coder.folding import getFolds
 
 import wx
@@ -24,6 +24,8 @@ class SourceTreePanel(wx.Panel):
         wx.Panel.__init__(self, parent, -1)
         self.parent = parent
         self.coder = frame
+
+        cs = ThemeMixin.appColors
 
         # get graphics for toolbars and tree items
         rc = self.coder.paths['icons']
