+++ psychopy/app/themes/__init__.py
@@ -1,1 +1,1 @@
-from ._themes import ThemeMixin
+from ._themes import ThemeMixin, cs_dark, cs_light
